Update
TCPFighter MMO
 - STL 사용 부분 Wrapping 하여 이후 컨테이너 교체에 용이하도록 처리
 - 프레임 드랍 발생시 로그 파일에 서버 상태 로깅하도록 처리
 - 불 필요한 헤더 파일 제거 (Util.h)
 - 라이브러리 참조 경로 수정
 - 환경 설정 파일 추가

+++ TCPFighter MMO/MMO_FightServer/MMO_FightServer/Object.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../Network/include/RingBuffer.h"
+#include "../Lib/Network/include/RingBuffer.h"
 
 struct SECTOR
 {
@@ -16,7 +16,7 @@ struct SECTOR_AROUND
 struct SESSION
 {
 	bool enable;
-	unsigned int socket;
+	DWORD64 socket;
 	wchar_t ip[16];
 	int port;
 	DWORD sessionID;
